fix: 스키마 코드 클린징
